Rewrite value convert to only have one value type
